tests/wpt/tests/css/css-ruby/inheritance.html

<!DOCTYPE html>
<html>
<head>
<meta charset="utf-8">
<title>Inheritance of CSS Ruby Layout properties</title>
<link rel="help" href="https://drafts.csswg.org/css-ruby-1/#property-index">
<meta name="assert" content="Properties inherit according to the spec.">
<meta name="assert" content="Properties have initial values according to the spec.">
<script src="/resources/testharness.js"></script>
<script src="/resources/testharnessreport.js"></script>
<script src="/css/support/inheritance-testcommon.js"></script>
</head>
<body>
<div id="container">
<div id="target"></div>
</div>
<script>
assert_inherited('ruby-align', 'space-around', 'center');
assert_inherited('ruby-position', 'alternate', 'under');
assert_inherited('ruby-merge', 'separate', 'merge');
assert_inherited('ruby-overhang', 'auto', 'none');
</script>
</body>
</html>
