third_party/blink/web_tests/external/wpt/css/css-ruby/inheritance-expected.txt

This is a testharness.js-based test.
[FAIL] Property ruby-position has initial value alternate
  assert_equals: expected "alternate" but got "over"
Harness: the test ran to completion.
